requirements.txt
python-dotenv
discogs-client
jinja2
requests
pylast==5.2.0